fix(deps): downgraded vault/sdk to v0.9.1
The unit tests failed because vault/sdk v0.9.2 introduced a new check
that I guess needs to be implemented - see [1]. By downgrading to
v0.9.1 this error can be fixed - but has to be addressed in the future.

This commit also fixes two linting errors and migrates the depguard
configuration to the new v2 schema. This is required because the
golangci-lint action bumped its dependency, see [2] and [3].

[1]: https://github.com/hashicorp/vault/issues/22173
[2]: https://github.com/golangci/golangci-lint/issues/3906
[3]: https://golangci-lint.run/usage/linters/#depguard

Signed-off-by: malmor <62105800+malmor@users.noreply.github.com>